Share the node check address table between server and app checks
The chain to known-address table moves from settings_chain into
primitives next to the node check types, so the Polkadot node probe in
check_node reuses the same address the server node checks use instead
of keeping its own constant.
