+++ core/bin/nodecheck/src/main.rs
@@ -9,7 +9,7 @@ use clap::Parser;
 use factory::new_provider;
 use gem_tracing::error_with_fields;
 use primitives::{Chain, NodeCheckProfile};
-use settings_chain::node_check_request;
+use primitives::node_check_request;
 
 use crate::service::NodeCheckService;
