+++ core/crates/gem_polkadot/src/constants.rs
@@ -1,5 +1,3 @@
-pub const TREASURY_ADDRESS: &str = "13UVJyLnbVp9RBZYFwFGyDvVd1y27Tt8tkntv6Q7JVPhFsTB";
-
 pub const TRANSACTION_TYPE_TRANSFER_KEEP_ALIVE: &str = "transferKeepAlive";
 pub const TRANSACTION_TYPE_TRANSFER_ALLOW_DEATH: &str = "transferAllowDeath";
 #[cfg(feature = "rpc")]
