+++ core/gemstone/src/services/chain/rules.rs
@@ -1,6 +1,6 @@
 use std::cmp::Reverse;
 
-use primitives::{AssetId, Chain, ChainAsset};
+use primitives::{AssetId, Chain, ChainAsset, NodeCheckProfile, NodeCheckRequest, node_check_request};
 
 pub fn chains_by_rank() -> Vec<Chain> {
     let mut chains = Chain::all();
@@ -32,9 +32,12 @@ pub fn is_valid_network_id(chain: Chain, network_id: &str) -> bool {
     chain.network_id() == network_id
 }
 
-pub fn node_verification_address(chain: Chain) -> Option<&'static str> {
+pub fn node_verification_address(chain: Chain) -> Option<String> {
     match chain {
-        Chain::Polkadot => Some(gem_polkadot::constants::TREASURY_ADDRESS),
+        Chain::Polkadot => match node_check_request(chain, NodeCheckProfile::Wallet) {
+            NodeCheckRequest::Wallet { address, .. } => Some(address),
+            NodeCheckRequest::Basic | NodeCheckRequest::Parser => None,
+        },
         _ => None,
     }
 }
